string fun

+++ kernel/types.h
@@ -2,7 +2,7 @@
  * @Author: Outsider
  * @Date: 2022-07-08 17:56:58
  * @LastEditors: Outsider
- * @LastEditTime: 2022-07-15 15:00:45
+ * @LastEditTime: 2022-07-19 17:07:31
  * @Description: 定义数据类型
  * @FilePath: /los/kernel/types.h
  */
@@ -15,8 +15,8 @@ typedef unsigned short uint16;
 typedef unsigned int uint32;
 typedef unsigned long long uint64;
 
-typedef unsigned int reg_t;     // 32bit register
-typedef unsigned long size_t;   // 32bit index
+typedef unsigned int reg_t;         // 32bit register
+typedef unsigned long long size_t;  // 64bit index
 
 typedef unsigned int pte_t;     // 32bit PTE
 typedef unsigned int addr_t;    // 32bit address;
